requirements.txt
distlib==0.3.7
filelock==3.13.1
greenlet==3.0.2
numpy==1.26.2
pandas==2.1.4
platformdirs==4.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.25.0
